float - int f3

+++ src/main/java/demo/BestClass.java
@@ -2,5 +2,6 @@ package demo;
 
 public class BestClass {
 
-    private float id;
+    //f3 float -> int
+    private int id;
 }
